wip: reorganize files
