+++ src/keyboard/keyMap.ts
@@ -15,18 +15,43 @@ export const defaultKeyMap: keyboardKey[] = [
   ...'ABCDEFGHIJKLMNOPQRSTUVWXYZ'
     .split('')
     .map(c => ({code: `Key${c}`, key: c, shiftKey: true})),
-    
+
   // alphanumeric block - functional
   {code: 'Space', key: ' '},
 
   {code: 'AltLeft', key: 'Alt', location: DOM_KEY_LOCATION.LEFT, keyCode: 18},
   {code: 'AltRight', key: 'Alt', location: DOM_KEY_LOCATION.RIGHT, keyCode: 18},
-  {code: 'ShiftLeft', key: 'Shift', location: DOM_KEY_LOCATION.LEFT, keyCode: 16},
-  {code: 'ShiftRight', key: 'Shift', location: DOM_KEY_LOCATION.RIGHT, keyCode: 16},
-  {code: 'ControlLeft', key: 'Control', location: DOM_KEY_LOCATION.LEFT, keyCode: 17},
-  {code: 'ControlRight', key: 'Control', location: DOM_KEY_LOCATION.RIGHT, keyCode: 17},
+  {
+    code: 'ShiftLeft',
+    key: 'Shift',
+    location: DOM_KEY_LOCATION.LEFT,
+    keyCode: 16,
+  },
+  {
+    code: 'ShiftRight',
+    key: 'Shift',
+    location: DOM_KEY_LOCATION.RIGHT,
+    keyCode: 16,
+  },
+  {
+    code: 'ControlLeft',
+    key: 'Control',
+    location: DOM_KEY_LOCATION.LEFT,
+    keyCode: 17,
+  },
+  {
+    code: 'ControlRight',
+    key: 'Control',
+    location: DOM_KEY_LOCATION.RIGHT,
+    keyCode: 17,
+  },
   {code: 'MetaLeft', key: 'Meta', location: DOM_KEY_LOCATION.LEFT, keyCode: 93},
-  {code: 'MetaRight', key: 'Meta', location: DOM_KEY_LOCATION.RIGHT, keyCode: 93},
+  {
+    code: 'MetaRight',
+    key: 'Meta',
+    location: DOM_KEY_LOCATION.RIGHT,
+    keyCode: 93,
+  },
 
   {code: 'OSLeft', key: 'OS', location: DOM_KEY_LOCATION.LEFT, keyCode: 91},
   {code: 'OSRight', key: 'OS', location: DOM_KEY_LOCATION.RIGHT, keyCode: 91},
@@ -36,17 +61,14 @@ export const defaultKeyMap: keyboardKey[] = [
   {code: 'Backspace', key: 'Backspace', keyCode: 8},
   {code: 'Enter', key: 'Enter', keyCode: 13},
 
-
   // function
   {code: 'Escape', key: 'Escape', keyCode: 27},
-
 
   // arrows
   {code: 'ArrowUp', key: 'ArrowUp', keyCode: 38},
   {code: 'ArrowDown', key: 'ArrowDown', keyCode: 40},
   {code: 'ArrowLeft', key: 'ArrowLeft', keyCode: 37},
   {code: 'ArrowRight', key: 'ArrowRight', keyCode: 39},
-
 
   // control pad
   {code: 'Home', key: 'Home', keyCode: 36},
