+++ src/keyboard/plugins/control.ts
