+++ src/keyboard/plugins/functional/calculateBackspaceValue.ts
