--- src/utils/click/index.ts
@@ -1,2 +1,0 @@
-export * from './getMouseEventOptions'
-export * from './isClickableInput'
