--- src/utils/focus/index.ts
@@ -1,3 +1,0 @@
-export * from './getActiveElement'
-export * from './isFocusable'
-export * from './selector'
