--- src/utils/misc/index.ts
@@ -1,5 +1,0 @@
-export * from './eventWrapper'
-export * from './isInstanceOfElement'
-export * from './isLabelWithInternallyDisabledControl'
-export * from './isVisible'
-export * from './isDisabled'
