+++ src/utils/misc/wait.ts
@@ -1,0 +1,3 @@
+export function wait(time?: number) {
+  return new Promise<void>(resolve => setTimeout(() => resolve(), time))
+}
